requirements.txt
appdirs==1.4.4
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
python-dotenv==0.19.0
regex==2021.8.21
requests==2.26.0
tomli==1.2.1
urllib3==1.26.6
Werkzeug==2.0.1
